chore: Move assignments tests to one file to remove duplicated fixtures
